fix: make use of port re-binding
previous implementation doesn't make use of ktors `reuseAddress` functionality since it manually checks if the port is in use. Often in these cases (if it was this process using the port) ktor can rebind the same port with no issues.

The new approach relies on letting Ktor tell us when it can't do this and handling the exception at that point instead of proactively throwing it.